Remove the crypto_ prefix from all crypto modules; move spake2p_test_vectors as a sub-module of spake2p
